chore: migrate infra to clickstream-base-lib
