+++ .projen/tasks.json
@@ -164,7 +164,12 @@
     },
     "pre-compile": {
       "name": "pre-compile",
-      "description": "Prepare the project for compilation"
+      "description": "Prepare the project for compilation",
+      "steps": [
+        {
+          "exec": "pnpm nx build @aws/clickstream-base-lib"
+        }
+      ]
     },
     "synth": {
       "name": "synth",
@@ -214,13 +219,13 @@
       },
       "steps": [
         {
-          "exec": "pnpm dlx npm-check-updates@16 --upgrade --target=minor --peer --dep=dev,peer,prod,optional --filter=@types/jest,esbuild,eslint-import-resolver-typescript,eslint-plugin-import,jest,projen,ts-jest,ts-node,typescript,p-limit"
+          "exec": "pnpm dlx npm-check-updates@16 --upgrade --target=minor --peer --dep=dev,peer,prod,optional --filter=@types/jest,esbuild,eslint-import-resolver-typescript,eslint-plugin-import,jest,projen,ts-jest,ts-node,typescript,@aws/clickstream-base-lib,p-limit"
         },
         {
           "exec": "pnpm i --no-frozen-lockfile"
         },
         {
-          "exec": "pnpm update @nrwl/devkit @nrwl/workspace @types/aws-lambda @types/jest @types/jsonpath-plus @types/jsonwebtoken @types/mock-fs @types/mustache @types/node-fetch @types/node @types/uuid @typescript-eslint/eslint-plugin @typescript-eslint/parser aws-cdk aws-sdk-client-mock-jest aws-sdk-client-mock esbuild eslint-import-resolver-typescript eslint-plugin-import eslint jest jest-junit mock-fs nx projen ts-jest ts-node typescript @aws-cdk/aws-glue-alpha @aws-cdk/aws-servicecatalogappregistry-alpha @aws-lambda-powertools/logger @aws-lambda-powertools/metrics @aws-sdk/client-acm @aws-sdk/client-cloudformation @aws-sdk/client-cloudwatch-events @aws-sdk/client-cloudwatch @aws-sdk/client-dynamodb @aws-sdk/client-ec2 @aws-sdk/client-ecs @aws-sdk/client-elastic-load-balancing-v2 @aws-sdk/client-emr-serverless @aws-sdk/client-glue @aws-sdk/client-iam @aws-sdk/client-kafka @aws-sdk/client-kafkaconnect @aws-sdk/client-lambda @aws-sdk/client-quicksight @aws-sdk/client-redshift-data @aws-sdk/client-redshift-serverless @aws-sdk/client-redshift @aws-sdk/client-route-53 @aws-sdk/client-s3 @aws-sdk/client-secrets-manager @aws-sdk/client-sfn @aws-sdk/client-sns @aws-sdk/client-sqs @aws-sdk/client-ssm @aws-sdk/client-sts @aws-sdk/credential-providers @aws-sdk/lib-dynamodb @aws-sdk/types @aws-sdk/util-dynamodb @smithy/node-http-handler @smithy/util-stream-node aws-cdk-lib cdk-bootstrapless-synthesizer cdk-nag cidr-block constructs cron-parser express-validator express fast-memoize hpagent i18next-fs-backend i18next json-difference jsonpath-plus jsonwebtoken jwks-rsa mustache node-cache node-fetch p-limit sql-formatter ts-json-object uuid"
+          "exec": "pnpm update @nrwl/devkit @nrwl/workspace @types/aws-lambda @types/jest @types/jsonpath-plus @types/jsonwebtoken @types/mock-fs @types/mustache @types/node-fetch @types/node @types/uuid @typescript-eslint/eslint-plugin @typescript-eslint/parser aws-cdk aws-sdk-client-mock-jest aws-sdk-client-mock esbuild eslint-import-resolver-typescript eslint-plugin-import eslint jest jest-junit mock-fs nx projen ts-jest ts-node typescript @aws-cdk/aws-glue-alpha @aws-cdk/aws-servicecatalogappregistry-alpha @aws-lambda-powertools/logger @aws-lambda-powertools/metrics @aws-sdk/client-acm @aws-sdk/client-cloudformation @aws-sdk/client-cloudwatch-events @aws-sdk/client-cloudwatch @aws-sdk/client-dynamodb @aws-sdk/client-ec2 @aws-sdk/client-ecs @aws-sdk/client-elastic-load-balancing-v2 @aws-sdk/client-emr-serverless @aws-sdk/client-glue @aws-sdk/client-iam @aws-sdk/client-kafka @aws-sdk/client-kafkaconnect @aws-sdk/client-lambda @aws-sdk/client-quicksight @aws-sdk/client-redshift-data @aws-sdk/client-redshift-serverless @aws-sdk/client-redshift @aws-sdk/client-route-53 @aws-sdk/client-s3 @aws-sdk/client-secrets-manager @aws-sdk/client-sfn @aws-sdk/client-sns @aws-sdk/client-sqs @aws-sdk/client-ssm @aws-sdk/client-sts @aws-sdk/credential-providers @aws-sdk/lib-dynamodb @aws-sdk/types @aws-sdk/util-dynamodb @aws/clickstream-base-lib @smithy/node-http-handler @smithy/util-stream-node aws-cdk-lib cdk-bootstrapless-synthesizer cdk-nag cidr-block constructs cron-parser express-validator express fast-memoize hpagent i18next-fs-backend i18next json-difference jsonpath-plus jsonwebtoken jwks-rsa mustache node-cache node-fetch p-limit sql-formatter ts-json-object uuid"
         },
         {
           "exec": "pnpm dlx projen"
