+++ package.json
@@ -89,6 +89,7 @@
     "@aws-sdk/lib-dynamodb": "^3.405.0",
     "@aws-sdk/types": "^3.398.0",
     "@aws-sdk/util-dynamodb": "^3.398.0",
+    "@aws/clickstream-base-lib": "workspace:*",
     "@smithy/node-http-handler": "^2.0.7",
     "@smithy/util-stream-node": "^2.0.7",
     "aws-cdk-lib": "^2.81.0",
