+++ sonar-project.properties
@@ -19,6 +19,7 @@ src/data-pipeline/spark-etl/*.gradle,\
 src/**/test/**/*,\
 frontend/**/test/**/*,\
 frontend/**/*.test.ts*,\
+src/base-lib/coverage/**/*,\
 *-ln.ts
 
 sonar.issue.ignore.multicriteria=e1
@@ -30,7 +31,7 @@ sonar.issue.ignore.multicriteria.e1.resourceKey=src/**/*.ts
 sonar.java.binaries=src/data-pipeline/spark-etl,src/data-pipeline/spark-etl/build/classes
 # Code coverage Specific Properties
 sonar.coverage.exclusions=deployment/**,examples/**,docs/**
-sonar.javascript.lcov.reportPaths=coverage/lcov.info,frontend/coverage/lcov.info
+sonar.javascript.lcov.reportPaths=coverage/lcov.info,frontend/coverage/lcov.info,src/base-lib/coverage/lcov.info
 sonar.coverage.jacoco.xmlReportPaths=src/data-pipeline/spark-etl/build/reports/jacoco/jacocoAggregatedReport/jacocoAggregatedReport.xml
 
 # Encoding of the source files
