+++ src/common/parameters.ts
@@ -12,11 +12,6 @@
  */
 
 import {
-  CfnParameter, CfnRule, Fn,
-} from 'aws-cdk-lib';
-
-import { Construct, IConstruct } from 'constructs';
-import {
   DOMAIN_NAME_PATTERN,
   HOST_ZONE_ID_PATTERN,
   PARAMETER_GROUP_LABEL_DOMAIN,
@@ -47,7 +42,12 @@ import {
   REDSHIFT_CLUSTER_IDENTIFIER_PATTERN,
   REDSHIFT_DB_USER_NAME_PATTERN,
   IAM_ROLE_PREFIX_PATTERN,
-} from './constant';
+} from '@aws/clickstream-base-lib';
+import {
+  CfnParameter, CfnRule, Fn,
+} from 'aws-cdk-lib';
+
+import { Construct, IConstruct } from 'constructs';
 import { KINESIS_MODE, REDSHIFT_MODE } from './model';
 
 export enum SubnetParameterType {
