+++ src/control-plane/backend/insert-admin-user.ts
@@ -11,13 +11,13 @@
  *  and limitations under the License.
  */
 
+import { DEFAULT_SOLUTION_OPERATOR } from '@aws/clickstream-base-lib';
 import { Duration } from 'aws-cdk-lib';
 import { Table } from 'aws-cdk-lib/aws-dynamodb';
 import { Effect, PolicyStatement } from 'aws-cdk-lib/aws-iam';
 import { RetentionDays } from 'aws-cdk-lib/aws-logs';
 import { AwsCustomResource, AwsCustomResourcePolicy, AwsSdkCall, PhysicalResourceId } from 'aws-cdk-lib/custom-resources';
 import { Construct } from 'constructs';
-import { DEFAULT_SOLUTION_OPERATOR } from '../../common/constant';
 
 export interface AddAdminUserProps {
   readonly uid: string;
