+++ src/data-pipeline/lambda/emr-serverless-app/index.ts
@@ -12,9 +12,9 @@
  */
 
 
+import { EMR_ARCHITECTURE_AUTO } from '@aws/clickstream-base-lib';
 import { EMRServerlessClient, CreateApplicationCommand, Architecture, CreateApplicationCommandInput, DeleteApplicationCommand } from '@aws-sdk/client-emr-serverless';
 import { CloudFormationCustomResourceEvent, Context } from 'aws-lambda';
-import { EMR_ARCHITECTURE_AUTO } from '../../../common/constant';
 import { logger } from '../../../common/powertools';
 import { putStringToS3, readS3ObjectAsJson } from '../../../common/s3';
 import { aws_sdk_client_common_config } from '../../../common/sdk-client-config';
