+++ src/ingestion-server-v2-stack.ts
@@ -11,6 +11,7 @@
  *  and limitations under the License.
  */
 
+import { OUTPUT_INGESTION_SERVER_DNS_SUFFIX, OUTPUT_INGESTION_SERVER_URL_SUFFIX } from '@aws/clickstream-base-lib';
 import {
   CfnCondition,
   CfnOutput,
@@ -28,7 +29,6 @@ import {
 import { Stream } from 'aws-cdk-lib/aws-kinesis';
 import { Construct } from 'constructs';
 import { RolePermissionBoundaryAspect } from './common/aspects';
-import { OUTPUT_INGESTION_SERVER_DNS_SUFFIX, OUTPUT_INGESTION_SERVER_URL_SUFFIX } from './common/constant';
 import { Parameters } from './common/parameters';
 import { SolutionInfo } from './common/solution-info';
 import { associateApplicationWithStack } from './common/stack';
