+++ src/kafka-s3-connector-stack.ts
@@ -11,6 +11,7 @@
  *  and limitations under the License.
  */
 
+import { DOMAIN_NAME_PATTERN } from '@aws/clickstream-base-lib';
 import { Aspects, CfnParameter, Fn, Stack, StackProps } from 'aws-cdk-lib';
 import { ISecurityGroup, SecurityGroup } from 'aws-cdk-lib/aws-ec2';
 import { Bucket, IBucket } from 'aws-cdk-lib/aws-s3';
@@ -18,7 +19,6 @@ import { NagSuppressions } from 'cdk-nag';
 import { Construct } from 'constructs';
 import { RolePermissionBoundaryAspect } from './common/aspects';
 import { addCfnNagForLogRetention, addCfnNagToStack, addCfnNagForCustomResourceProvider, ruleToSuppressRolePolicyWithWildcardResources, ruleToSuppressRolePolicyWithHighSPCM } from './common/cfn-nag';
-import { DOMAIN_NAME_PATTERN } from './common/constant';
 import { Parameters } from './common/parameters';
 import { SolutionInfo } from './common/solution-info';
 import { associateApplicationWithStack } from './common/stack';
