+++ src/metrics-stack.ts
@@ -11,6 +11,7 @@
  *  and limitations under the License.
  */
 
+import { EMAIL_PATTERN, OUTPUT_METRICS_OBSERVABILITY_DASHBOARD_NAME, OUTPUT_METRICS_SNS_TOPIC_ARN_NAME, PROJECT_ID_PATTERN } from '@aws/clickstream-base-lib';
 import { Aspects, CfnOutput, CfnParameter, Fn, Stack, StackProps } from 'aws-cdk-lib';
 import { PolicyStatement, ServicePrincipal } from 'aws-cdk-lib/aws-iam';
 import { Key } from 'aws-cdk-lib/aws-kms';
@@ -18,7 +19,6 @@ import { Topic } from 'aws-cdk-lib/aws-sns';
 import { Construct } from 'constructs';
 import { RolePermissionBoundaryAspect } from './common/aspects';
 import { addCfnNagForCustomResourceProvider, addCfnNagForLogRetention } from './common/cfn-nag';
-import { EMAIL_PATTERN, OUTPUT_METRICS_OBSERVABILITY_DASHBOARD_NAME, OUTPUT_METRICS_SNS_TOPIC_ARN_NAME, PROJECT_ID_PATTERN } from './common/constant';
 import { Parameters } from './common/parameters';
 import { SolutionInfo } from './common/solution-info';
 import { associateApplicationWithStack } from './common/stack';
