+++ src/metrics/util.ts
@@ -11,13 +11,13 @@
  *  and limitations under the License.
  */
 
+import { ALARM_NAME_PREFIX, METRICS_PARAMETER_PATH_PREFIX } from '@aws/clickstream-base-lib';
 import { GetParametersByPathCommand, Parameter, SSMClient } from '@aws-sdk/client-ssm';
 import { CfnResource, Stack } from 'aws-cdk-lib';
 import { Alarm } from 'aws-cdk-lib/aws-cloudwatch';
 import { Construct } from 'constructs';
 import { WIDGETS_HEIGHT } from './settings';
 import { addCfnNagSuppressRules } from '../common/cfn-nag';
-import { ALARM_NAME_PREFIX, METRICS_PARAMETER_PATH_PREFIX } from '../common/constant';
 import { logger } from '../common/powertools';
 import { getShortIdOfStack } from '../common/stack';
