+++ test/analytics/analytics-on-redshift/lambda/load-data/has-more-job-new.test.ts
@@ -11,21 +11,20 @@
  *  and limitations under the License.
  */
 
-
 process.env.REDSHIFT_ODS_TABLE_NAME = 'test_me_table';
 process.env.ODS_EVENT_BUCKET_PREFIX = 'project1/test/test_me_table/';
 
+import { PARTITION_APP } from '@aws/clickstream-base-lib';
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
 import { QueryCommand } from '@aws-sdk/lib-dynamodb';
 import { mockClient } from 'aws-sdk-client-mock';
 import 'aws-sdk-client-mock-jest';
 import { handler } from '../../../../../src/analytics/lambdas/load-data-workflow/has-more-job-new';
 import { JobStatus } from '../../../../../src/analytics/private/constant';
-import { PARTITION_APP } from '../../../../../src/common/constant';
 import { getMockContext } from '../../../../common/lambda-context';
 
-const ddbClientMock = mockClient(DynamoDBClient);
 
+const ddbClientMock = mockClient(DynamoDBClient);
 const context = getMockContext();
 
 beforeEach(() => {
