+++ test/analytics/analytics-on-redshift/lambda/load-data/skip-running-workflow.test.ts
@@ -11,6 +11,7 @@
  *  and limitations under the License.
  */
 
+import { PARTITION_APP } from '@aws/clickstream-base-lib';
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
 import { SFNClient, ListExecutionsCommand, ExecutionStatus } from '@aws-sdk/client-sfn';
 import { QueryCommand } from '@aws-sdk/lib-dynamodb';
@@ -18,9 +19,7 @@ import { mockClient } from 'aws-sdk-client-mock';
 import 'aws-sdk-client-mock-jest';
 import { handler } from '../../../../../src/analytics/lambdas/load-data-workflow/skip-running-workflow';
 import { JobStatus, REDSHIFT_TABLE_NAMES } from '../../../../../src/analytics/private/constant';
-import { PARTITION_APP } from '../../../../../src/common/constant';
 import { getMockContext } from '../../../../common/lambda-context';
-
 
 const ddbClientMock = mockClient(DynamoDBClient);
 const snfClientMock = mockClient(SFNClient);
