+++ test/control-plane/cloudfront-control-plane-stack.test.ts
@@ -11,10 +11,10 @@
  *  and limitations under the License.
  */
 
+import { OUTPUT_CONTROL_PLANE_URL, OUTPUT_CONTROL_PLANE_BUCKET } from '@aws/clickstream-base-lib';
 import { Capture, Match, Template } from 'aws-cdk-lib/assertions';
 import { findResourcesName } from './test-utils';
 import { CloudFrontControlPlaneStack } from '../../src/cloudfront-control-plane-stack';
-import { OUTPUT_CONTROL_PLANE_URL, OUTPUT_CONTROL_PLANE_BUCKET } from '../../src/common/constant';
 import { TestApp, removeFolder } from '../common/jest';
 import { CFN_FN } from '../constants';
 import { findFirstResourceByKeyPrefix } from '../utils';
