+++ test/reporting/quicksight/data-reporting-quicksight-stack.test.ts
@@ -11,9 +11,9 @@
  *  and limitations under the License.
  */
 
+import { OUTPUT_REPORTING_QUICKSIGHT_DASHBOARDS, OUTPUT_REPORTING_QUICKSIGHT_DATA_SOURCE_ARN } from '@aws/clickstream-base-lib';
 import { App } from 'aws-cdk-lib';
 import { Match, Template } from 'aws-cdk-lib/assertions';
-import { OUTPUT_REPORTING_QUICKSIGHT_DASHBOARDS, OUTPUT_REPORTING_QUICKSIGHT_DATA_SOURCE_ARN } from '../../../src/common/constant';
 import { DataReportingQuickSightStack } from '../../../src/data-reporting-quicksight-stack';
 
 describe('DataReportingQuickSightStack parameter test', () => {
